new _zip_folder_to_file function
