+++ fractal_server/zip_tools.py
@@ -1,23 +1,50 @@
 import os
+import shutil
 from io import BytesIO
+from pathlib import Path
+from typing import Generator
 from zipfile import ZIP_DEFLATED
 from zipfile import ZipFile
 
 
-def _zip_folder_to_byte_stream(*, folder: str) -> BytesIO:
+def _zip_folder_to_byte_stream(*, folder: str) -> Generator:
     """
     Get byte stream with the zipped log folder of a job.
 
     Args:
         folder: the folder to zip
     """
+    zip_file = Path(f"{folder}.zip")
 
-    byte_stream = BytesIO()
-    with ZipFile(byte_stream, mode="w", compression=ZIP_DEFLATED) as zipfile:
-        for root, dirs, files in os.walk(folder):
-            for file in files:
-                file_path = os.path.join(root, file)
-                archive_path = os.path.relpath(file_path, folder)
-                zipfile.write(file_path, archive_path)
+    if os.path.exists(zip_file):
 
-    return byte_stream
+        def iterfile():
+            """
+            https://fastapi.tiangolo.com/advanced/
+            custom-response/#using-streamingresponse-with-file-like-objects
+            """
+            with open(zip_file, mode="rb") as file_like:
+                yield from file_like
+
+        return iterfile()
+
+    else:
+
+        byte_stream = BytesIO()
+        with ZipFile(
+            byte_stream, mode="w", compression=ZIP_DEFLATED
+        ) as zipfile:
+            for root, dirs, files in os.walk(folder):
+                for file in files:
+                    file_path = os.path.join(root, file)
+                    archive_path = os.path.relpath(file_path, folder)
+                    zipfile.write(file_path, archive_path)
+        return iter([byte_stream.getvalue()])
+
+
+def _zip_folder_to_file(*, folder: str) -> None:
+    shutil.make_archive(
+        base_name=f"{folder}_tmp", format="zip", root_dir=Path(folder)
+    )
+    shutil.move(f"{folder}_tmp.zip", f"{folder}.zip")
+    shutil.rmtree(folder)
